Navigations via swipe gestures should also skip back/forward list items without a user gesture https://bugs.webkit.org/show_bug.cgi?id=241932
Reviewed by Geoffrey Garen.

Navigations via swipe gestures should also skip back/forward list items without a user gesture.
Previously, with Bug 241885, it only impacted navigations via the back/forward buttons.

* Source/WebKit/UIProcess/ViewGestureController.cpp:
(WebKit::ViewGestureController::startSwipeGesture):
* Source/WebKit/UIProcess/WebBackForwardList.cpp:
(WebKit::itemSkippingBackForwardItemsAddedByJSWithoutUserGesture):
(WebKit::WebBackForwardList::goBackItemSkippingItemsWithoutUserGesture const):
(WebKit::WebBackForwardList::goForwardItemSkippingItemsWithoutUserGesture const):
* Source/WebKit/UIProcess/WebBackForwardList.h:
* Source/WebKit/UIProcess/WebPageProxy.cpp:
(WebKit::WebPageProxy::goForward):
(WebKit::WebPageProxy::goBack):
(WebKit::itemSkippingBackForwardItemsAddedByJSWithoutUserGesture): Deleted.

Canonical link: https://commits.webkit.org/251837@main
